server/requirements.txt
asgiref==3.4.1
Django==2.2.17
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.6.1
Markdown==3.3.4
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.5.0
